refactor: service logic 분리
